script/requirements.txt
requests
beautifulsoup4
google-cloud-texttospeech
google-cloud-speech
pandas
opencv-python
numpy
Pillow